requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
anyio==4.11.0
asgiref==3.9.0
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.2
boto3==1.34.144
botocore==1.34.162
cachetools==5.5.2
celery==5.3.6
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==0.51.0
django-channels==0.7.0
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-rest-framework==0.1.0
django-storages==1.14.6
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
exponent_server_sdk==2.2.0
frozenlist==1.7.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googlemaps==4.10.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.3.7
MarkupSafe==3.0.3
msgpack==1.1.2
multidict==6.6.4
numpy==2.4.1
oauthlib==3.3.1
packaging==25.0
pandas==3.0.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
razorpay==2.0.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
twilio==9.7.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.1
